고다은 : 새로운 게시판 (리퀘스트) 추가 작업
● 리퀘스트 게시판

○ 폼 입력
- (뷰) book_request_form.jsp : 로그인한 사용자만 작성할 수 있도록, 포워딩으로 받은 세션 아이디 사용 (if문)
- (서비스) RequestFormToDBService : 세션아이디가 null이 아닐때만 적재 가능

○ 폼 수정
- (뷰) book_request_detail.jsp
 - 수정 버튼 폼으로 변경 (히든으로 reqnum을 넘김)
- (서비스) RequestDetailUpdateFormService
 - reqnum이 null로 올 수도 있으므로 if로 처리해줌
 - 세션 아이디 바인딩
- (뷰) book_request_detail_update_form.jsp
 - 다른 사람이 작성한 수정폼에 들어오면 해당 리뷰 리스트로 리다이렉트

○ 폼 삭제
- (서비스) RequestDeleteService
 - 폼아이디와 세션 아이디를 비교해서 같은 경우만 삭제 되도록

○ 구매허가 버튼
- (서비스) RequestPermissionService
 - 누를 때마다 책을 book 테이블에 적재하는 것을 막기 위해, book 테이블에 해당 도서가 없는 경우에만 적재 가능하도록.

○ 조회
- (뷰) book_request_detail.jsp
 - 주소창에 그냥 requestDetail.do 쳤을때 커버하는 코드 추가

+++ electronic library/src/main/java/kr/co/ict/servlet/service/request/RequestDeleteService.java
@@ -5,6 +5,7 @@ import java.io.IOException;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import kr.co.ict.RequestDAO;
 
@@ -13,11 +14,22 @@ public class RequestDeleteService implements IRequestService {
 	@Override
 	public void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
-		int reqNum = Integer.parseInt(request.getParameter("reqnum"));
+		String strNum = request.getParameter("reqnum");
+		int reqNum = 0;
+		if(strNum != null) {
+			reqNum = Integer.parseInt(strNum);
+		}
+		String fId = request.getParameter("fid");
 		
-		// 다오 메서드
-		RequestDAO dao = RequestDAO.getInstance();
-		dao.deleteBookRequest(reqNum);
+		// 세션
+		HttpSession session = request.getSession();
+		String sId = (String)session.getAttribute("sId");
+		
+		if(fId.equals(sId)) {
+			// 다오 메서드
+			RequestDAO dao = RequestDAO.getInstance();
+			dao.deleteBookRequest(reqNum);
+		}
 	}
 
 }
